Moved food code into two seperate classes
