+++ UnityProject/Group8/Assets/Scenes/Level.unity
@@ -925,6 +925,8 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   currentPrawn: {fileID: 0}
   currentDumpling: {fileID: 0}
+  currentSalmon: {fileID: 0}
+  currentNigri: {fileID: 0}
 --- !u!4 &1353850398
 Transform:
   m_ObjectHideFlags: 0
