+++ UnityProject/Group8/Assets/Scripts/Food.cs
@@ -8,6 +8,9 @@ public class Food : MonoBehaviour
     public GameObject food;
     public LevelManager level;
 
+    [Header("Components")]
+    public SpriteRenderer sprite;
+
     [Range(0.0f, 6.0f)]
     public float speed;
 
@@ -24,31 +27,5 @@ public class Food : MonoBehaviour
     void OnBecameInvisible()
     {
         // TODO: Call destroy food function from level manager
-    }
-}
-
-// TODO: May need a separate script file for each food class.
-// TODO: Implement Prawn class here
-// TODO: Implement Dumpling class here
-
-// Salmon inherits all of food behavior from the base class 
-public class Salmon : Food
-{
-    // TODO: We could set speed here
-
-    private void OnBecameInvisible() 
-    {
-        level.GetComponent<LevelManager>().DestroySalmon();
-    }
-}
-
-// Nigri inherits all of food behavior from the base class 
-public class Nigri : Food
-{
-    // TODO: We could set speed here
-
-    private void OnBecameInvisible()
-    {
-        level.GetComponent<LevelManager>().DestroyNigri();
     }
 }
