+++ UnityProject/Group8/Assets/Scripts/New Prefab.prefab.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: ec7113c58f6690c46af10f81733e8eef
+timeCreated: 1520777229
+licenseType: Free
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
