+++ UnityProject/Group8/Assets/Scripts/Nigri.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 4dafe14309422da499b2f129f6116f0c
+timeCreated: 1520776709
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
