+++ UnityProject/Group8/Assets/Scripts/Salmon.cs
@@ -1,0 +1,14 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+// Salmon inherits all of food behavior from the base class 
+public class Salmon : Food
+{
+    // TODO: We could set speed here
+
+    private void OnBecameInvisible()
+    {
+        level.GetComponent<LevelManager>().DestroySalmon();
+    }
+}
