+++ UnityProject/Group8/Assets/Scripts/Salmon.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: bab8f7b05623c6741b3c936751bdc025
+timeCreated: 1520776700
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
